BISOL Energija App/requirements.txt
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
anyio @ file:///io/perseverance-python-buildout/croot/anyio_1731700736723/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///io/buildout/croot/certifi_1735842370795/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi @ file:///io/buildout/croot/fastapi_1735845387481/work
fastapi-cli==0.0.7
greenlet @ file:///croot/greenlet_1733860072744/work
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
numexpr @ file:///io/perseverance-python-buildout/croot/numexpr_1731701292781/work
numpy @ file:///croot/numpy_and_numpy_base_1736500434252/work/dist/numpy-2.2.1-cp313-cp313-linux_x86_64.whl#sha256=0bf4a7adf3743a7d50ab1bbcd49e5fb31ec820e590c41eeab0d32a0235e0029d
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp313-cp313-linux_x86_64.whl#sha256=248baa16b8f7b48e3a9c0e641fba9a72bce9c95c0dddf7c9a7547016b61fe3fb
pyasn1==0.6.1
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments==2.19.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil @ file:///work/perseverance-python-buildout/croot/python-dateutil_1728385355105/work
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz @ file:///work/perseverance-python-buildout/croot/pytz_1728386814080/work
PyYAML==6.0.2
requests @ file:///croot/requests_1730999120400/work
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==72.1.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///work/perseverance-python-buildout/croot/sniffio_1728386342155/work
SQLAlchemy @ file:///croot/sqlalchemy_1736565039440/work
sqlmodel==0.0.22
starlette @ file:///io/perseverance-python-buildout/croot/starlette-recipe_1731704598518/work
typer==0.15.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.44.0
